requirements_enhanced.txt
# 基础依赖
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0
scikit-learn>=1.3.0
scikit-image>=0.21.0

# 系统监控和性能
psutil>=5.9.0
memory-profiler>=0.60.0

# 并行处理
joblib>=1.3.0
multiprocessing-logging>=0.3.0

# 深度学习框架（可选）
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
keras>=2.13.0

# GPU加速（可选）
cupy-cuda11x>=12.0.0; platform_system != "Windows"
cupy-cuda12x>=12.0.0; platform_system != "Windows"

# 图像处理增强
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
imgaug>=0.4.0

# 数据科学和可视化
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# 配置和日志
pyyaml>=6.0
python-json-logger>=2.0.0
colorlog>=6.7.0

# 测试和基准测试
pytest>=7.4.0
pytest-benchmark>=4.0.0
pytest-cov>=4.1.0

# 类型检查
mypy>=1.5.0
types-PyYAML>=6.0.0

# 代码质量
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# 文档生成
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# 开发工具
jupyter>=1.0.0
ipython>=8.14.0
tqdm>=4.65.0

# 网络和API
requests>=2.31.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# 数据库（可选）
sqlalchemy>=2.0.0
alembic>=1.11.0

# 缓存和队列
redis>=4.6.0
celery>=5.3.0

# 安全和加密
cryptography>=41.0.0
python-dotenv>=1.0.0

# 文件处理
pathlib2>=2.3.7; python_version < "3.4"
watchdog>=3.0.0

# 数学和科学计算
scipy>=1.11.0
pandas>=2.0.0

# 图像格式支持
libvips>=8.14.0
pyvips>=2.2.0

# 机器学习工具
xgboost>=1.7.0
lightgbm>=4.0.0

# 模型优化
onnx>=1.14.0
onnxruntime>=1.15.0

# 模型部署
torchserve>=0.8.0
tensorflow-serving>=2.13.0

# 监控和指标
prometheus-client>=0.17.0
grafana-api>=1.0.0

# 错误追踪
sentry-sdk>=1.28.0

# 性能分析
py-spy>=0.3.14
memory-profiler>=0.60.0

# 容器化支持
docker>=6.1.0

# 云服务集成
boto3>=1.28.0
google-cloud-storage>=2.10.0
azure-storage-blob>=12.17.0 